:recycle: PostViewModel 생성 , 함수 이동
